+++ compose/src/main/java/com/yourssu/design/system/compose/atom/ListItem.kt
@@ -14,7 +14,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.yourssu.design.system.compose.R
 import com.yourssu.design.system.compose.YdsTheme
-import com.yourssu.design.system.compose.foundation.YdsIcon
+import com.yourssu.design.system.compose.base.YdsIcon
 import com.yourssu.design.system.compose.base.noRippleClickable
 
 @Composable
